+++ core/comps/contacts/srvc-contact-model.js
@@ -114,16 +114,18 @@ angular.module('cmContacts')
             };
 
             this.delete = function(withoutModal){
-                return withoutModal
+                return (function(){
+                    return withoutModal
                     ? $q.when()
                     : cmModal.confirm({
                         title: 'CONTACT.MODAL.DELETE.HEADER',
                         text: 'CONTACT.MODAL.DELETE.TEXT'
-                }).then(function() {
-                    return cmContactsAdapter
-                        .deleteContact(self.id)
+                    })
+                }()).then(function(){
+                    return cmContactsAdapter.deleteContact(self.id)
                 }).then(function(){
                     self.trigger('deleted:finished',self);
+                    return $q.when();
                 });
             };