+++ test/e2e/routes/settings/check-key-maximum.1.spec.js
@@ -1,5 +1,6 @@
 var config = require("../../config/specs.js")
-var util = require("../../../lib/e2e/cmTestUtil.js")
+var util = require("../../cmUtil.js")
+
 describe('Check key maximum: ',function(){
     var ptor = util.getPtorInstance(),
         testUser